requirements.txt
pypng==0.0.20
pyqrcode==1.2.1
reportlab==3.5.9
